+++ randomizer-data/src/main/java/io/github/astrapi69/random/RandomCharacters.java
@@ -1,0 +1,110 @@
+/**
+ * The MIT License
+ * <p>
+ * Copyright (C) 2015 Asterios Raptis
+ * <p>
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ * <p>
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * <p>
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.random;
+
+/**
+ * The enum {@link RandomCharacters}
+ */
+public enum RandomCharacters
+{
+
+	/** The lowcase. */
+	lowcase(RandomCharacters.LOWCASECHARS),
+
+	/** The lowcase with numbers. */
+	lowcaseWithNumbers(RandomCharacters.LOWCASECHARS + RandomCharacters.NUMBERS),
+
+	/** The lowcase with numbers and special. */
+	lowcaseWithNumbersAndSpecial(
+		RandomCharacters.LOWCASECHARS + RandomCharacters.NUMBERS + RandomCharacters.SPECIALCHARS),
+
+	/** The lowcase with uppercase and numbers. */
+	lowcaseWithUppercaseAndNumbers(RandomCharacters.LOWCASECHARS + RandomCharacters.LOWCASECHARS
+		.toUpperCase() + RandomCharacters.NUMBERS),
+
+	/** The lowcase with uppercase and numbers and special. */
+	lowcaseWithUppercaseAndNumbersAndSpecial(
+		RandomCharacters.LOWCASECHARS + RandomCharacters.LOWCASECHARS
+			.toUpperCase() + RandomCharacters.NUMBERS + RandomCharacters.SPECIALCHARS),
+
+	/** The numbers. */
+	numbers(RandomCharacters.NUMBERS),
+
+	/** The special. */
+	special(RandomCharacters.SPECIALCHARS),
+
+	/** The uppercase. */
+	uppercase(RandomCharacters.LOWCASECHARS.toUpperCase()),
+
+	/** The uppercase with numbers. */
+	uppercaseWithNumbers(RandomCharacters.LOWCASECHARS.toUpperCase() + RandomCharacters.NUMBERS),
+
+	/** The uppercase. */
+	escapeSequences(RandomCharacters.ESCAPE_SEQUENCES),
+
+	/** The escape sequences and white space. */
+	escapeSequencesWithWhitespace(RandomCharacters.ESCAPE_SEQUENCES_CHARS),
+
+	/** The uppercase with numbers and special. */
+	uppercaseWithNumbersAndSpecial(RandomCharacters.LOWCASECHARS
+		.toUpperCase() + RandomCharacters.NUMBERS + RandomCharacters.SPECIALCHARS);
+
+	/** The alphabet-characters in lower case. */
+	public static final String LOWCASECHARS = "abcdefghijklmnopqrstuvwxyz";
+
+	/** All digits. */
+	public static final String NUMBERS = "0123456789";
+
+	/** All special characters */
+	public static final String SPECIALCHARS = "#@$%^&*?!";
+
+	/** The white space char */
+	public static final String WHITE_SPACE_CHAR = " ";
+
+	/** All escape sequences characters */
+	public static final String ESCAPE_SEQUENCES = "\t\b\n\r\f\'\"\\";
+
+	/** All escape sequences and white space characters */
+	public static final String ESCAPE_SEQUENCES_CHARS = ESCAPE_SEQUENCES + WHITE_SPACE_CHAR;
+
+	/** The allowed characters. */
+	private final String characters;
+
+	/**
+	 * Instantiates a new {@link RandomCharacters}
+	 *
+	 * @param characters
+	 *            the allowed characters
+	 */
+	private RandomCharacters(String characters)
+	{
+		this.characters = characters;
+	}
+
+	public String getCharacters()
+	{
+		return characters;
+	}
+}